requirements.txt
neo4j>=5.0,<6.0
httpx>=0.27,<1.0
pandas>=2.0,<3.0
python-dotenv>=1.0,<2.0
lxml>=5.0,<6.0
pytest>=8.0,<9.0
ruff>=0.4,<1.0
